+++ packages/desktop/.storybook/__mocks__/defaultEndpoints.js
@@ -1,48 +1,63 @@
+import { Chance } from "chance";
+
+const chance = new Chance();
+
+function fakeCountry() {
+  const id = chance.guid();
+  const country = chance.country({ full: true });
+  return { id, name: country };
+}
+
+function fakeCity(locationGroup) {
+  const id = chance.guid();
+  const city = chance.city();
+  const availability = chance.pickone(['AVAILABLE', 'RESERVED', 'OCCUPIED']);
+  return { id, name: city, locationGroup, availability };
+}
+
+function fakeLabTestLaboratory() {
+  const id = chance.guid();
+  const name = `${chance.first()} ${chance.animal().toLowerCase()} laboratory`
+  return { id, name };
+}
+
+function fakeLabTestPriority() {
+  const id = chance.guid();
+  const adverb = chance.pickone(['Total', 'Slight', 'Critical', 'Ultimate', 'Bare', 'Unknown', 'Unknowable', 'Unbelievable']);
+  const noun = chance.pickone(['importance', 'urgency', 'criticality', 'priority', 'seriousness', 'danger']);
+  return { id, name: `${adverb} ${noun}` };
+}
+
+function fakeLabTestCategory() {
+  const id = chance.guid();
+  const area = chance.pickone(['Blood', 'Urine', 'Stool', 'Hormone', 'Immunology', 'Radiology', 'Other'])
+  const testSynonym = chance.pickone(['trial', 'test', 'analysis', 'examination', 'investigation', 'screening', 'diagnosis', 'assessment', 'analysis', 'study'])
+  return { id, name: `${area} ${testSynonym}` };
+}
+
+export const fakeLocations = [];
+
+for (let i = 0; i < 10; i++) {
+  const country = fakeCountry();
+  fakeLocations.push(country);
+
+  for (let j = 0; j < 20; j++) {
+    const city = fakeCity(country);
+    fakeLocations.push(city);
+  }
+}
+
 export const defaultEndpoints = {
-  'suggestions/labTestLaboratory/all': () => [
-    {
-      id: '1',
-      name: 'ABC Medical Center',
-    },
-    {
-      id: '2',
-      name: 'XYZ Diagnostics',
-    },
-    {
-      id: '3',
-      name: 'City Hospital Lab',
-    },
-    {
-      id: '4',
-      name: 'Sunrise Diagnostic Center',
-    },
-  ],
-  'suggestions/labTestPriority/all': () => [
-    {
-      id: '1',
-      name: 'Chill',
-    },
-    {
-      id: '2',
-      name: 'Very Important',
-    },
-    {
-      id: '3',
-      name: 'Critical',
-    }
-  ],
-  'suggestions/labTestCategory/all': () => [
-    {
-      id: '1',
-      name: 'Blood Tests',
-    },
-    {
-      id: '2',
-      name: 'Urine Tests',
-    },
-    {
-      id: '3',
-      name: 'Radiology Tests',
-    },
-  ],
+  'suggestions/locationGroup/all': () => {
+    return fakeLocations;
+  },
+  'suggestions/location': () => {
+    return fakeLocations;
+  },
+  'suggestions/location/:id': (data, id) => {
+    return fakeLocations.find(x => x.id === id);
+  },
+  'suggestions/labTestLaboratory/all': () => Array.from({ length: 10 }, fakeLabTestLaboratory),
+  'suggestions/labTestPriority/all': () => Array.from({ length: 10 }, fakeLabTestPriority),
+  'suggestions/labTestCategory/all': () => Array.from({ length: 10 }, fakeLabTestCategory)
 };
