+++ packages/desktop/stories/locationField.stories.js
@@ -1,10 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
-import Chance from 'chance';
 import { storiesOf } from '@storybook/react';
 import { Typography, Box } from '@material-ui/core';
 import { Form, Field, LocationField } from '../app/components';
-import { MockedApi } from './utils/mockedApi';
+import { fakeLocations } from '../.storybook/__mocks__/defaultEndpoints';
 
 /**
  * TODO: Semi-broken from changes to suggester logic
@@ -30,51 +29,7 @@ const OneColumn = styled.div`
   grid-row-gap: 24px;
 `;
 
-const chance = new Chance();
-
-function fakeCountry() {
-  const id = chance.guid();
-  const country = chance.country({ full: true });
-  return { id, name: country };
-}
-
-function fakeCity(locationGroup) {
-  const id = chance.guid();
-  const city = chance.city();
-  const availability = chance.pickone(['AVAILABLE', 'RESERVED', 'OCCUPIED']);
-  return { id, name: city, locationGroup, availability };
-}
-
-const fakeLocations = [];
-
-for (let i = 0; i < 10; i++) {
-  const country = fakeCountry();
-  fakeLocations.push(country);
-
-  for (let j = 0; j < 20; j++) {
-    const city = fakeCity(country);
-    fakeLocations.push(city);
-  }
-}
-
-const endpoints = {
-  'suggestions/locationGroup/all': () => {
-    return fakeLocations;
-  },
-  'suggestions/location': () => {
-    return fakeLocations;
-  },
-  'suggestions/location/:id': (data, id) => {
-    return fakeLocations.find(x => x.id === id);
-  },
-};
-
 storiesOf('LocationField', module)
-  .addDecorator(Story => (
-    <MockedApi endpoints={endpoints}>
-      <Story />
-    </MockedApi>
-  ))
   .add('One Column', () => {
     return (
       <Form
